add struct philo
add smart sleep in many usleep

bug pointeur begin_time last_meal

+++ philo/src/time.c
@@ -6,16 +6,16 @@
 /*   By: jvigny <jvigny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/13 20:13:31 by jvigny            #+#    #+#             */
-/*   Updated: 2023/02/16 17:11:03 by jvigny           ###   ########.fr       */
+/*   Updated: 2023/02/16 19:34:24 by jvigny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-int	ft_time(struct timeval *begin)
+long	ft_time(struct timeval *begin)
 {
 	struct timeval time;
-	int				milli;
+	long			milli;
 
 	gettimeofday(&time, NULL);
 	// printf("time : %ld\n", (time).tv_usec / 1000 + (time).tv_sec * 1000);
